shiny/driver/mtldriver: add rgba->bgra conversion for m1s
It appears that newer macs have deprecated the ability to render metal windows as rgba, which was the default behavior on older macs even when bgra was requested

This is a confusing statement and is not backed up by anything I can find online, but this patch on m1 fixes the display such that the r and b channels are flipped.

Candidate for backporting to v3.